staging

+++ DLWMLS/utils.py
@@ -11,8 +11,7 @@ def prepare_data_folder(folder_path: str) -> None:
     if not os.path.exists(folder_path):
         os.makedirs(folder_path)
 
-
-def rename_and_copy_files(src_folder: str, des_folder: str, suffix="_DLWMLS.nii.gz") -> Tuple[dict, dict]:
+def rename_and_copy_files(src_folder: str, des_folder: str, suffix: str) -> Tuple[dict, dict]:
     """
     Input:
          src_folder: a user input folder, name could be anything, will be convert into nnUnet
@@ -25,16 +24,21 @@ def rename_and_copy_files(src_folder: str, des_folder: str, suffix="_DLWMLS.nii.
          rename_back_dict:  a dictionary will be use to mapping backto the original name
 
     """
+    if not os.path.exists(src_folder):
+        raise FileNotFoundError(f"Source folder '{src_folder}' does not exist.")
+    if not os.path.exists(des_folder):
+        raise FileNotFoundError(f"Source folder '{des_folder}' does not exist.")
+
     files = os.listdir(src_folder)
     rename_dict = {}
     rename_back_dict = {}
 
     for idx, filename in enumerate(files):
         old_name = os.path.join(src_folder, filename)
-        if not os.path.isfile(old_name):
+        if not os.path.isfile(old_name):  # We only want files!
             continue
-        rename_file = f"case_{idx: 04d}_0000.nii.gz"
-        rename_back = f"case_{idx: 04d}.nii.gz"
+        rename_file = f"case_{idx:03d}_0000.nii.gz"
+        rename_back = f"case_{idx:03d}.nii.gz"
         new_name = os.path.join(des_folder, rename_file)
         try:
             shutil.copy2(old_name, new_name)
@@ -42,5 +46,6 @@ def rename_and_copy_files(src_folder: str, des_folder: str, suffix="_DLWMLS.nii.
             rename_back_dict[rename_back] = filename.split(".nii")[0] + suffix
         except Exception as e:
             print(f"Error copying file '{filename}' to '{new_name}': {e}")
-
+    print(rename_back_dict)
+    
     return rename_dict, rename_back_dict
